pr comments resolution
