+++ apps/quick-learn-backend/project.json
@@ -5,20 +5,24 @@
   "projectType": "application",
   "tags": [],
   "targets": {
-    // "build": {
-    //   "executor": "@nx/webpack:webpack",
-    //   "outputs": ["{options.outputPath}"],
-    //   "options": {
-    //     "target": "node",
-    //     "compiler": "tsc",
-    //     "outputPath": "dist/apps/quick-learn-backend",
-    //     "main": "apps/quick-learn-backend/src/main.ts",
-    //     "tsConfig": "apps/quick-learn-backend/tsconfig.app.json",
-    //     "assets": ["apps/quick-learn-backend/src/assets"],
-    //     "isolatedConfig": true,
-    //     "webpackConfig": "apps/quick-learn-backend/webpack.config.js"
-    //   }
-    // },
+    "build": {
+      "executor": "@nx/webpack:webpack",
+      "outputs": [
+        "{options.outputPath}"
+      ],
+      "options": {
+        "target": "node",
+        "compiler": "tsc",
+        "outputPath": "dist/apps/quick-learn-backend",
+        "main": "apps/quick-learn-backend/src/main.ts",
+        "tsConfig": "apps/quick-learn-backend/tsconfig.app.json",
+        "assets": [
+          "apps/quick-learn-backend/src/assets"
+        ],
+        "isolatedConfig": true,
+        "webpackConfig": "apps/quick-learn-backend/webpack.config.js"
+      }
+    },
     "serve": {
       "executor": "@nx/js:node",
       "defaultConfiguration": "development",
@@ -86,7 +90,9 @@
     },
     "lint": {
       "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
+      "outputs": [
+        "{options.outputFile}"
+      ],
       "options": {
         "lintFilePatterns": [
           "apps/quick-learn-backend/src/**/*.{ts,tsx,js,jsx}"
