+++ apps/quick-learn-backend/src/routes/team/team.controller.ts
@@ -6,6 +6,7 @@ import { SuccessResponse } from '@src/common/dto';
 import { TeamService } from './team.service';
 
 @ApiTags('Team')
+@UseGuards(JwtAuthGuard)
 // using the global prefix from main file (api) and putting versioning here as v1 /api/v1/team
 @Controller({
   version: '1',
@@ -14,7 +15,6 @@ import { TeamService } from './team.service';
 export class TeamController {
   constructor(private readonly teamService: TeamService) {}
 
-  @UseGuards(JwtAuthGuard)
   @Post()
   @ApiOperation({ summary: 'adding team name' })
   async create(@Body() createTeamDto: CreateTeamDto): Promise<SuccessResponse> {
@@ -23,7 +23,6 @@ export class TeamController {
     return new SuccessResponse('Successfully added Team', team);
   }
 
-  @UseGuards(JwtAuthGuard)
   @Get()
   @ApiOperation({ summary: 'get all team names' })
   async findAll() {
