+++ apps/quick-learn-frontend/src/apiServices/accountService.ts
@@ -11,15 +11,6 @@ import {
   TTeam,
 } from '@src/shared/types/accountTypes';
 
-// export const getUserProfile = async (): Promise<
-//   AxiosSuccessResponse<TUserProfileType>
-// > => {
-//   const response = await axiosInstance.get<
-//     AxiosSuccessResponse<TUserProfileType>
-//   >(userApiEnum.GET_USER_PROFILE);
-//   return response.data;
-// };
-
 export const addSkill = async (data: TSkill): Promise<AxiosSuccessResponse> => {
   const response = await axiosInstance.post<AxiosSuccessResponse>(
     accountApiEnum.SKILLS,
